[DEV-3450] filter login ip (#1844)
* [DEV-3450] filter login ip

* [DEV-3450] add await

* [DEV-3450] Refactoring

* [DEV-3450] Refactoring 2
