+++ src/shared/auth/ip-country.guard.ts
@@ -1,5 +1,6 @@
 import { CanActivate, ExecutionContext, ForbiddenException, Injectable } from '@nestjs/common';
 import { getClientIp } from '@supercharge/request-ip';
+import { Config } from 'src/config/config';
 import { IpLogService } from '../models/ip-log/ip-log.service';
 
 @Injectable()
@@ -11,6 +12,9 @@ export class IpCountryGuard implements CanActivate {
 
     const ipLog = await this.ipLogService.create(ip, req.url, req.body?.address ?? req.user.address);
     if (!ipLog.result) throw new ForbiddenException('The country of IP address is not allowed');
+
+    if (req.body?.region && !Config.loginCountries[req.body?.region]?.includes(ipLog.country))
+      throw new ForbiddenException('The country of IP address is not allowed');
 
     return ipLog.result;
   }
