+++ src/subdomains/generic/user/models/auth/dto/auth-credentials.dto.ts
@@ -1,9 +1,9 @@
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
-import { IsNotEmpty, IsOptional, IsString, Matches, ValidateIf } from 'class-validator';
+import { IsInt, IsNotEmpty, IsOptional, IsString, Matches, ValidateIf } from 'class-validator';
 import { GetConfig } from 'src/config/config';
 import { CryptoService } from 'src/integration/blockchain/shared/services/crypto.service';
 
-export class AuthCredentialsDto {
+export class SignInDto {
   @ApiProperty()
   @IsNotEmpty()
   @IsString()
@@ -21,8 +21,7 @@ export class AuthCredentialsDto {
   @IsString()
   @Matches(GetConfig().formats.key)
   @ValidateIf(
-    (dto: AuthCredentialsDto) =>
-      CryptoService.isArweaveAddress(dto.address) || CryptoService.isCardanoAddress(dto.address),
+    (dto: SignInDto) => CryptoService.isArweaveAddress(dto.address) || CryptoService.isCardanoAddress(dto.address),
   )
   key?: string;
 
@@ -35,4 +34,53 @@ export class AuthCredentialsDto {
   @IsOptional()
   @IsString()
   specialCode?: string;
+
+  @ApiPropertyOptional({ description: 'IP region filter' })
+  @IsOptional()
+  @IsString()
+  region?: string;
+}
+
+export class OptionalSignUpDto {
+  @ApiPropertyOptional()
+  @IsOptional()
+  @IsString()
+  @Matches(GetConfig().formats.ref)
+  usedRef?: string;
+
+  @IsOptional()
+  @IsInt()
+  walletId?: number;
+
+  @ApiPropertyOptional()
+  @IsOptional()
+  @IsString()
+  wallet?: string;
+
+  @ApiPropertyOptional({ deprecated: true, description: 'This field is deprecated, use "specialCode" instead.' })
+  @IsOptional()
+  @IsString()
+  discountCode?: string;
+
+  @ApiPropertyOptional({ description: 'Special code' })
+  @IsOptional()
+  @IsString()
+  specialCode?: string;
+}
+
+export class SignUpDto extends SignInDto {
+  @ApiPropertyOptional()
+  @IsOptional()
+  @IsString()
+  @Matches(GetConfig().formats.ref)
+  usedRef?: string;
+
+  @IsOptional()
+  @IsInt()
+  walletId?: number;
+
+  @ApiPropertyOptional()
+  @IsOptional()
+  @IsString()
+  wallet?: string;
 }
