+++ src/subdomains/generic/user/models/user/dto/alby.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsNotEmpty, IsString } from 'class-validator';
-import { OptionalSignUpDto } from './create-user.dto';
+import { OptionalSignUpDto } from '../../auth/dto/auth-credentials.dto';
 
 export class AlbySignupDto extends OptionalSignUpDto {
   @ApiProperty()
